Add convolutional, max pooling and reshape layers.
The code now includes the implementation of convolutional, max pooling, and reshape layers essential for deep onvolutional Neural Networks (CNNs). These new capabilities allow the construction of more complex and powerful network architectures. In addition, the Dense layer has been updated with the ability to compute the derivatives of weights and biases for the backpropagation process. The Activation functions utilized in the Convolutional layer have been provided as well.